# backend/routers/projects.py
from __future__ import annotations

import json
import logging
import os

from fastapi import APIRouter, HTTPException

from backend.agents.ingestion import ingest_corpus
from backend.api_context import _load_json, _safe_id
from backend.paths import CORPUS, PROJECTS_DIR

router = APIRouter()
log = logging.getLogger("pramaan.api")

# ── Pipeline info endpoint ───────────────────────────────────────────

@router.get("/pipeline")
def pipeline_info():
    return {
        "name": "Pramaan Compliance Reasoning Graph (with bounded retrieval + critique cycles)",
        "framework": "LangGraph",
        "llm_footprint": "single LLM reasoning core at node_reconcile; other nodes are deterministic-first",
        "nodes": [
            {"id": "ingest", "node": "Ingestion", "kind": "deterministic",
             "description": "Document intake, parsing, normalization"},
            {"id": "load_standards", "node": "Standards Loader", "kind": "deterministic",
             "description": "Load governing standards corpus"},
            {"id": "validate", "node": "Validation Gate", "kind": "deterministic",
             "description": "Check spec+submittal exist; conditional routing"},
            {"id": "reconcile", "node": "Reconciliation", "kind": "llm_core",
             "description": "Cross-document deviation reasoning"},
            {"id": "retrieve", "node": "Standards Retrieval Tool", "kind": "deterministic",
             "description": "Fetches a cited standard absent from context; loops back to reconcile (tool-call cycle)"},
            {"id": "critique", "node": "Self-Critique Gate", "kind": "deterministic_by_default",
             "description": "Verifies its own findings; loops back to reconcile on a failed self-check (reflexion)"},
            {"id": "cx_predict", "node": "Cx Predictor", "kind": "rule_graph_first",
             "description": "Map deviations to commissioning tests"},
            {"id": "format_output", "node": "Output Formatter", "kind": "deterministic",
             "description": "Enrich and structure findings"},
        ],
        "edges": [
            ["ingest", "load_standards"],
            ["load_standards", "validate"],
            {"from": "validate", "to": ["reconcile", "format_output"], "type": "conditional",
             "condition": "route_after_validate: skip reconciliation if spec or submittal missing"},
            ["reconcile", "retrieve"],
            {"from": "retrieve", "to": ["reconcile", "critique"], "type": "conditional", "cycle": True,
             "condition": "route_after_retrieve: loop back to reconcile after fetching a missing cited standard "
                          "(bounded by PRAMAAN_MAX_RETRIEVALS), else proceed"},
            {"from": "critique", "to": ["reconcile", "cx_predict"], "type": "conditional", "cycle": True,
             "condition": "route_after_critique: loop back to reconcile on a failed self-check "
                          "(bounded by PRAMAAN_MAX_REVISIONS), else proceed"},
            ["cx_predict", "format_output"],
        ],
        "services": [
            {"id": "extraction", "service": "Extraction service", "description": "Raw document to structured triples"},
            {"id": "rfi_copilot", "service": "RFI Copilot",
             "description": "RAG over project corpus with prior-RFI matching"},
        ],
    }


@router.get("/corpus/doc/{doc_type}/{system_id}")
def corpus_doc(doc_type: str, system_id: str):
    if doc_type not in ("specs", "submittals"):
        raise HTTPException(400, "doc_type must be 'specs' or 'submittals'")
    system_id = _safe_id(system_id, "system_id")
    path = CORPUS / doc_type / f"{system_id}.md"
    if not path.exists():
        raise HTTPException(404, f"Document not found: {doc_type}/{system_id}")
    return {"system": system_id, "doc_type": doc_type, "text": path.read_text(encoding="utf-8")}


@router.get("/corpus/stats")
def corpus_stats():
    result = ingest_corpus()
    standards_dir = CORPUS / "standards"
    total_lines = 0
    if standards_dir.exists():
        for f in standards_dir.glob("*.md"):
            total_lines += f.read_text().count("\n")
    return {
        "total_systems": result.get("total_systems", len(result.get("systems", []))),
        "total_standards": result.get("total_standards", len(result.get("standards", []))),
        "total_documents": result.get("total_documents", 0),
        "standards_lines": total_lines,
    }


# ── Multi-project endpoints ─────────────────────────────────────────

def _project_summary(gt: dict, project_id: str) -> dict:
    proj = gt.get("project", {})
    devs = gt.get("seeded_deviations", [])
    return {
        "id": project_id,
        "name": proj.get("name", project_id),
        "tier": proj.get("tier", ""),
        "location": proj.get("location", ""),
        "capacity_mw": proj.get("capacity_mw", 0),
        "deviations": len(devs),
        "systems": proj.get("total_systems", 0),
    }


@router.get("/projects")
def list_projects():
    projects = []
    gt = _load_json("ground_truth.json")
    if gt:
        projects.append(_project_summary(gt, "meghdoot"))

    if PROJECTS_DIR.exists():
        for p in sorted(PROJECTS_DIR.iterdir()):
            gt_path = p / "ground_truth.json"
            if p.is_dir() and gt_path.exists():
                pgt = json.loads(gt_path.read_text())
                projects.append(_project_summary(pgt, p.name))

    return {"projects": projects, "count": len(projects)}


def _deviation_summary(deviations: list[dict]) -> dict:
    lead_times = [item["lead_time_weeks"] for item in deviations if item.get("lead_time_weeks") is not None]
    return {
        "count": len(deviations),
        "critical": sum(1 for item in deviations if item.get("severity") == "Critical"),
        "major": sum(1 for item in deviations if item.get("severity") == "Major"),
        "total_lead_time_weeks": sum(lead_times),
        "max_lead_time_weeks": max(lead_times, default=0),
    }


@router.get("/projects/{project_id}")
def project_detail(project_id: str):
    project_id = _safe_id(project_id, "project_id")
    ppath = CORPUS if project_id == "meghdoot" else PROJECTS_DIR / project_id

    gt_path = ppath / "ground_truth.json"
    if not gt_path.exists():
        raise HTTPException(404, f"Project '{project_id}' not found")

    gt = json.loads(gt_path.read_text())
    cx_path = ppath / "commissioning" / "cx_plan.json"
    cx = json.loads(cx_path.read_text()) if cx_path.exists() else {}

    devs = gt.get("seeded_deviations", [])

    return {
        "project": gt.get("project", {}),
        "deviations": devs,
        "deviation_summary": _deviation_summary(devs),
        "cx_plan": cx,
        "true_negative_systems": gt.get("true_negative_systems", []),
    }


# ── PS4 capability layers: schedule risk · supply chain · unified graph ──────
# Each is gated by an env flag (default on) and returns {"available": false}
# instead of 404/500 when its data file or flag is absent, so the existing 19
# sections and the 315-test baseline are never affected.

# Per-project memo cache for the PS4 layers. The analysis is a pure function of
# static on-disk JSON that never changes during a demo, so the first hit pays the
# Monte-Carlo cost (~1-2s on a free-tier box) and every subsequent hit is instant —
# and immune to concurrent-judge-click stalls.
_PS4_CACHE: dict = {}


def _project_dir(project_id: str):
    project_id = _safe_id(project_id, "project_id")
    return CORPUS if project_id == "meghdoot" else PROJECTS_DIR / project_id


def _load_project_file(project_id: str, name: str):
    p = _project_dir(project_id) / name
    if not p.exists():
        return None
    try:
        return json.loads(p.read_text(encoding="utf-8"))
    except Exception:
        return None


@router.get("/projects/{project_id}/schedule")
def project_schedule(project_id: str):
    if os.getenv("PRAMAAN_SCHEDULE", "1") == "0":
        return {"available": False}
    cache_key = ("schedule", project_id)
    if cache_key in _PS4_CACHE:
        return _PS4_CACHE[cache_key]
    try:
        sch = _load_project_file(project_id, "schedule.json")
        if not sch:
            return {"available": False}
        from backend.agents import schedule_risk
        gt = _load_project_file(project_id, "ground_truth.json") or {}
        sch = {**sch, "risks": schedule_risk.derive_risks(sch, gt.get("seeded_deviations", []))}
        analysis = schedule_risk.analyze_schedule(sch, n=5000, seed=42)
        analysis["available"] = True
        analysis["narrative"] = schedule_risk.narrate(analysis)
        _PS4_CACHE[cache_key] = analysis
        return analysis
    except Exception as exc:
        log.warning("Schedule analysis failed for %s: %s", project_id, exc)
        return {"available": False, "error": str(exc)}


@router.get("/projects/{project_id}/supply-chain")
def project_supply_chain(project_id: str):
    if os.getenv("PRAMAAN_SUPPLY", "1") == "0":
        return {"available": False}
    try:
        sup = _load_project_file(project_id, "supply_chain.json")
        if not sup:
            return {"available": False}
        from backend.agents import supply_chain
        analysis = supply_chain.analyze_supply_chain(sup)
        analysis["available"] = True
        analysis["narrative"] = supply_chain.narrate(analysis)
        return analysis
    except Exception as exc:
        log.warning("Supply-chain analysis failed for %s: %s", project_id, exc)
        return {"available": False, "error": str(exc)}


def _assemble_project_graph(project_id: str):
    from backend.agents import project_graph
    gt = _load_project_file(project_id, "ground_truth.json")
    if not gt:
        return None, None
    g = project_graph.assemble(
        gt.get("seeded_deviations", []),
        cx_plan=_load_project_file(project_id, "commissioning/cx_plan.json"),
        supply_chain=_load_project_file(project_id, "supply_chain.json"),
        schedule=_load_project_file(project_id, "schedule.json"),
    )
    return project_graph, g


@router.get("/projects/{project_id}/graph")
def project_graph_view(project_id: str):
    if os.getenv("PRAMAAN_GRAPH", "1") == "0":
        return {"available": False}
    try:
        pg, g = _assemble_project_graph(project_id)
        if g is None:
            return {"available": False}
        return {"available": True, "stats": pg.graph_stats(g), "graph": pg.as_graph(g)}
    except Exception as exc:
        log.warning("Project graph failed for %s: %s", project_id, exc)
        return {"available": False, "error": str(exc)}


@router.get("/projects/{project_id}/blast-radius/{dev_id}")
def project_blast_radius(project_id: str, dev_id: str):
    if os.getenv("PRAMAAN_GRAPH", "1") == "0":
        return {"available": False}
    try:
        dev_id = _safe_id(dev_id, "dev_id")
        pg, g = _assemble_project_graph(project_id)
        if g is None:
            return {"available": False}
        br = pg.blast_radius(g, dev_id)
        if br is None:
            return {"available": False}
        return {"available": True, **br}
    except Exception as exc:
        log.warning("Blast radius failed for %s/%s: %s", project_id, dev_id, exc)
        return {"available": False, "error": str(exc)}


@router.get("/projects/{project_id}/remediation/{dev_id}")
def project_remediation(project_id: str, dev_id: str, cost_per_week_lakh: float = 200.0):
    """What-if remediation simulator: how the schedule slip (and cost) of a
    deviation changes with the week it is caught — flat-zero until the cliff,
    then climbing one-for-one. Deterministic; the LLM never touches it."""
    if os.getenv("PRAMAAN_GRAPH", "1") == "0":
        return {"available": False}
    try:
        dev_id = _safe_id(dev_id, "dev_id")
        pg, g = _assemble_project_graph(project_id)
        if g is None:
            return {"available": False}
        sim = pg.simulate_remediation(g, dev_id, cost_per_week_lakh=cost_per_week_lakh)
        if sim is None:
            return {"available": False}
        return {"available": True, **sim}
    except Exception as exc:
        log.warning("Remediation sim failed for %s/%s: %s", project_id, dev_id, exc)
        return {"available": False, "error": str(exc)}


@router.get("/projects/eval/aggregate")
def projects_eval_aggregate():
    try:
        return _compute_projects_aggregate()
    except Exception as exc:
        log.warning("projects aggregate unavailable: %s", exc)
        return {"aggregate": {}, "per_project": {}, "error": "aggregate temporarily unavailable"}


def _compute_projects_aggregate():
    from eval.multi_project_eval import aggregate, run_all
    results = run_all()
    agg = aggregate(results)
    per_project = {
        pid: {
            "name": r["name"],
            "tier": r["tier"],
            "location": r["location"],
            "capacity_mw": r["capacity_mw"],
            "deviations": r["deviations"],
            "precision": round(r["scores"]["precision"], 3),
            "recall": round(r["scores"]["recall"], 3),
            "f1": round(r["scores"]["f1"], 3),
            "cx_accuracy": round(r["scores"]["cx_prediction_accuracy"], 3),
            "total_lead_weeks": r["scores"]["total_lead_time_weeks"],
        }
        for pid, r in results.items()
    }
    return {"aggregate": agg, "per_project": per_project}