# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Spec-to-Site Deviation Sentinel for hyperscale data-centre EPC delivery",
    "title": "Pramaan — EPC Deviation Intelligence",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/corpus/doc/{doc_type}/{system_id}": {
      "get": {
        "operationId": "v1_corpus_doc",
        "parameters": [
          {
            "in": "path",
            "name": "doc_type",
            "required": true,
            "schema": {
              "title": "Doc Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "system_id",
            "required": true,
            "schema": {
              "title": "System Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Corpus Doc",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}": {
      "get": {
        "operationId": "v1_project_detail",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Project Detail",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/projects/{project_id}/blast-radius/{dev_id}": {
      "get": {
        "operationId": "v1_project_blast_radius",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dev_id",
            "required": true,
            "schema": {
              "title": "Dev Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Project Blast Radius",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/projects/{project_id}/graph": {
      "get": {
        "operationId": "v1_project_graph_view",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Project Graph View",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/projects/{project_id}/remediation/{dev_id}": {
      "get": {
        "description": "What-if remediation simulator: how the schedule slip (and cost) of a\ndeviation changes with the week it is caught — flat-zero until the cliff,\nthen climbing one-for-one. Deterministic; the LLM never touches it.",
        "operationId": "v1_project_remediation",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dev_id",
            "required": true,
            "schema": {
              "title": "Dev Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cost_per_week_lakh",
            "required": false,
            "schema": {
              "default": 200.0,
              "title": "Cost Per Week Lakh",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Project Remediation",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/projects/{project_id}/schedule": {
      "get": {
        "operationId": "v1_project_schedule",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Project Schedule",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/projects/{project_id}/supply-chain": {
      "get": {
        "operationId": "v1_project_supply_chain",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Project Supply Chain",
        "tags": [
          "v1"
        ]
      }
    },
...
  }
}